Update Docker configuration to use nginx instead of OpenResty (#801)
Replaced the OpenResty image with the nginx:alpine image in both local
Docker Compose files. Removed the command for installing LuaRocks and
running OpenResty, simplifying the setup for local development.

Thanks to @sharang-d for the ping about the issue with luarocks.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Chores**
* Updated the Docker Compose configuration to use the standard nginx
image instead of OpenResty for local development environments.
* Simplified the nginx service startup by removing custom commands and
package installations.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->