Add column `external_id` to table dividends (#972)
Follows https://github.com/antiwork/flexile/pull/767

I noticed we expose the ID when reviewing that PR, so I had made a note
to fix it.

We need to add this column and populate it. Then we can make all
dividend/dividend_computation pages use the external ID in the URL.